add role is now responsive, name changes
